requirements.txt
streamlit==1.32.0
torch==2.0.1
torchvision==0.15.2
pillow==10.2.0
numpy==1.26.0